Extract the common work with the Kafka Streams config into a utility.
